Add an environment and a resource for a node network with a floating IP
The environment can be used by setting:

cluster_environment_nodenet: "{{ role_path }}/files/environments/nodenet-w-fip.yaml"

If this environment is used, then cluster_params.cluster_net[0] should contain
a "floating_net" field, which is the name or ID of the network on which to
create floating IPs.

This change also adds support for a "security_groups" field in
cluster_params.cluster_net[0] for both the existing nodenet-1.yaml environment
and the new nodenet-w-fip.yaml environment.